Add requirements and update setup instructions
