+++ .gitignore
@@ -1,3 +1,3 @@
 # Ignore asset files
-assets/*.png
-assets/*.wav
+assets/images/*.png
+assets/sounds/*.wav
